moved cif handeling to another file
